digital_shop/requirements.txt
Babel==2.7.0
boto3==1.9.224
botocore==1.12.224
certifi==2019.6.16
chardet==3.0.4
construct==2.5.3
defusedxml==0.6.0
Django==2.0.13
django-extra-views==0.11.0
django-haystack==2.8.1
django-localflavor==2.2
django-oscar==2.0
django-oscar-paypal==1.0.0
django-phonenumber-field==2.0.1
django-storages==1.7.1
django-tables2==1.21.2
django-treebeard==4.3
django-widget-tweaks==1.4.5
docutils==0.15.2
factory-boy==2.12.0
Faker==2.0.1
future==0.17.1
idna==2.8
jmespath==0.9.4
mock==2.0.0
oauthlib==3.1.0
pbr==5.4.3
pefile==2019.4.18
phonenumbers==8.10.18
Pillow==6.1.0
purl==1.5
PyJWT==1.7.1
python-dateutil==2.8.0
python-ptrace==0.9.4
python-stdnum==1.11
python3-openid==3.1.0
pytz==2019.2
requests==2.22.0
requests-oauthlib==1.2.0
s3transfer==0.2.1
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sorl-thumbnail==12.4.1
sqlparse==0.3.0
text-unidecode==1.2
Unidecode==1.0.23
urllib3==1.25.3
